Support custom workfile and idempotent left over processing
The custom_workfile property allows to specify the filename of a
'path_*.txt'-like file, which is expected to be found under
<data_dir>/<country_name>/<custom_workfile>.

Previously, the left_over property could be used to point to an
alternative workfile. Now, left_over is used to only process cities
which haven't been processed yet, i.e. where no <city>_<left_over>.csv
file exist.

For example, to run feature-engineering in an idempotent manner only for
cities for which features have not been calculated yet, specify
bld_d_fts as left_over property in the config.yml. It will be checked if
a <path>/<city>_bld_d_fts.csv file already exists before starting
feature-engineering for a particular city.