Change log directory structure
Every run of a slurm pipeline job creates its own log dir containing the
*work.json files. All individual task logs are stored in a subdirectory
to improve clarity.

The job name %x is removed from the individual task logs as it is
already indicated by the file path.

The work dir name doesn't need to be prefixed by a uuid because its
parent directory, the log dir, is already uniquely named due to the
timestamp.