Remove outdated array conf generation
The slurm task array was determined based on the number of lines of the
workfile. In the past, for txt workfiles this was appropriate, for json
encoded workfiles this is incorrect.

Remove the logic that builds the array conf based on the workfile and
expect the array definition to be explicitely specified as an argument.