Process workfile concurrently if job array is not specified
For each param combination in the workfile, start the Python script and
immediately background it, so that the sbatch-workfile.sh script is not
blocked but can process the next param combination.

Note: this potentially puts high computational load on the node. If the
resource allocation for each iteration is high, use a slurm task array
instead to distribute the computational load.