Support dynamic generation of job params based on param grid config
Introduce option in job config to specify the path to a params generator
config file. Each combination (multi-dimensional Cartesian product) of
the attributes in the param_generator_file is determined and a params
list is generated (equivalent to the content of a param_file).

An example param_generator_file could look like:
param_1: [value_1, value_2, value_3]
param_2: [constant]

which would generate the following param list:
- param_1: value_1
  param_2: constant
- param_1: value_2
  param_2: constant
- param_1: value_3
  param_2: constant

Note:
The param_files and param_generator_file options are mutually exclusive,
but one is always required.