Improve displaying of logs
Integrate questionary package to allow users to interactively select the
work package for which the logs shall be displayed.

Further, refactor stderr and stdout cmd options and remove --index
option in favor of allowing users to specify the index after the job
name, e.g. `--job=some-name.0` to view the logs of the first work
package.