Refactor slack notification logic
* Only post a single message per job to a channel.
* Instead of sending status updates in the thread, update main message
  itself to reduce verboseness.
* Send logs and a summary of Slurm statuses for failing work packages to
  thread.
* Mention @channel once if some work package fails.
* Mention @channel when job finishes.

Note: Use workaround to get the Slack channel id required for updating a
message. Slack API
[postMessage](https://api.slack.com/methods/chat.postMessage) allows to
send a message given only a channel name, but returns the actual channel
id. We update the slack_channel instance variable with the retrieved id
in order to use Slack API
[update](https://api.slack.com/methods/chat.update), which doesn't
support a channel name. This workaround works as long as the bot first
sends a message before it updates one.