Ensure code blocks in long Slack messages are preserved
The Slack API method chat.postMessage splits messages in multiple parts
if longer than 4,000 characters. This commit ensures that code blocks
are closed and opened correctly when a long message is split to preserve
the Markdown-style formatting of the message.

Limitation: Currently only supported for sending new messages, not for
updating existing ones (e.g. the work_status will not be formatted
correctly).

Add unit test to test that each message part contains an equal number of
"```" and no information is accidentially truncated.